requirements.txt
alpha_vantage==2.3.1
keras==2.11.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
python-decouple==3.8
scikit_learn==1.2.2
tensorflow==2.11.0
tensorflow_intel==2.11.0
vectorbt==0.24.5
yfinance==0.2.12
